Remove uncalled code (base generator amongst others) (#7682)
