+++ rake_tasks/crazy_fun/mappings/export.rb
@@ -2,6 +2,4 @@ require 'rake_tasks/crazy_fun/mappings/common'
 
 require_relative 'export_mappings'
 require_relative 'export/check_preconditions'
-require_relative 'export/export_tasks'
-require_relative 'export/create_task'
 require_relative 'export/add_dependencies'
