+++ rake_tasks/crazy_fun/mappings/python/py_task.rb
@@ -1,11 +1,9 @@
 module Python
   class PyTask < Tasks
-    def get_resources(browser, args)
+    def get_resources(_browser, args)
       resources = []
       resources.concat(args[:resources]) if args[:resources]
-      browser_specific_resources = SeleniumRake::Browsers::BROWSERS[browser][:python][:resources]
-      resources.concat(browser_specific_resources) if browser_specific_resources
-      return resources
+      resources
     end
   end
 end
